+++ src/spi/exclusive_device.rs
@@ -1,9 +1,4 @@
-use core::convert::Infallible;
-
-use embedded_hal::{
-    blocking::spi::{Operation, Transactional, Transfer, Write, WriteIter},
-    spi::FullDuplex,
-};
+use embedded_hal::spi::{blocking::SpiDevice, ErrorKind, ErrorType};
 
 use crate::spi::SpiConfig;
 
@@ -36,85 +31,25 @@ where
     }
 }
 
-impl<SPI, PINS> FullDuplex<u8> for SpiExclusiveDevice<SPI, PINS>
-where
-    SPI: SpiX,
-    PINS: Pins<SPI>,
-{
-    type Error = Infallible;
-
-    fn read(&mut self) -> nb::Result<u8, Infallible> {
-        self.bus.read()
-    }
-
-    fn send(&mut self, byte: u8) -> nb::Result<(), Infallible> {
-        self.bus.send(byte)
-    }
+impl<SPI, PINS> ErrorType for SpiExclusiveDevice<SPI, PINS> {
+    type Error = ErrorKind;
 }
 
-impl<SPI, PINS> Transfer<u8> for SpiExclusiveDevice<SPI, PINS>
+impl<SPI, PINS> SpiDevice for SpiExclusiveDevice<SPI, PINS>
 where
     SPI: SpiX,
     PINS: Pins<SPI>,
 {
-    type Error = Infallible;
+    type Bus = SpiBus<SPI, PINS>;
 
-    fn transfer<'w>(&mut self, words: &'w mut [u8]) -> Result<&'w [u8], Self::Error> {
+    fn transaction<R>(
+        &mut self,
+        f: impl FnOnce(&mut Self::Bus) -> Result<R, <Self::Bus as ErrorType>::Error>,
+    ) -> Result<R, Self::Error> {
         self.bus.start_frame();
-        let result = self.bus.transfer(words);
+        let result = f(&mut self.bus)?;
         self.bus.end_frame();
 
-        result
-    }
-}
-
-impl<SPI, PINS> Write<u8> for SpiExclusiveDevice<SPI, PINS>
-where
-    SPI: SpiX,
-    PINS: Pins<SPI>,
-{
-    type Error = Infallible;
-
-    fn write(&mut self, words: &[u8]) -> Result<(), Self::Error> {
-        self.bus.start_frame();
-        let result = self.bus.write(words);
-        self.bus.end_frame();
-
-        result
-    }
-}
-
-impl<SPI, PINS> WriteIter<u8> for SpiExclusiveDevice<SPI, PINS>
-where
-    SPI: SpiX,
-    PINS: Pins<SPI>,
-{
-    type Error = Infallible;
-
-    fn write_iter<WI>(&mut self, words: WI) -> Result<(), Self::Error>
-    where
-        WI: IntoIterator<Item = u8>,
-    {
-        self.bus.start_frame();
-        let result = self.bus.write_iter(words);
-        self.bus.end_frame();
-
-        result
-    }
-}
-
-impl<SPI, PINS> Transactional<u8> for SpiExclusiveDevice<SPI, PINS>
-where
-    SPI: SpiX,
-    PINS: Pins<SPI>,
-{
-    type Error = Infallible;
-
-    fn exec<'op>(&mut self, operations: &mut [Operation<'op, u8>]) -> Result<(), Infallible> {
-        self.bus.start_frame();
-        let result = self.bus.exec(operations);
-        self.bus.end_frame();
-
-        result
+        Ok(result)
     }
 }